+++ modules/server/src/data-access/repository-services/player.ts
@@ -1,0 +1,10 @@
+import { PlayerModel } from '@dirtleague/common';
+import Repository from './repository';
+
+class PlayerRepository extends Repository {
+  update = async (model: PlayerModel): Promise<void> => {
+    await this.context.profiles.update(model);
+  };
+}
+
+export default PlayerRepository;
